requirements.txt
# Trade Genie - Python Dependencies
# Install with: pip install -r requirements.txt

# ---- Web Framework ----
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.12

# ---- LLM Providers ----
openai==1.55.3
anthropic==0.39.0
google-generativeai==0.8.3

# ---- Data Sources ----
requests==2.32.3
feedparser==6.0.11          # RSS/Atom feeds
praw==7.8.1                 # Reddit API

# ---- Scheduling ----
apscheduler==3.10.4

# ---- Config ----
python-dotenv==1.0.1

# ---- Utilities ----
python-dateutil==2.9.0
